fix the test code which had errors
